refactor(money): dinero en entero CLP en toda la cadena (Fase 1)

+++ metaproduct/models.py
@@ -7,7 +7,7 @@ Origen = (
 # Create your models here.
 class Material(models.Model):
     name = models.CharField(max_length=255)
-    cost = models.DecimalField(max_digits=7, decimal_places=2)
+    cost = models.PositiveIntegerField()  # entero CLP
     def __str__(self):
         return self.name
